backend/requirements.txt
flask
flask-cors
flask_socketio
numpy
tensorflow
pymilvus
openai
boto3
ollama
litellm
python-dotenv
scikit-learn
torch>=1.7.0
torchvision>=0.8.1
opencv-python
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.41.0
matplotlib>=3.2.2
seaborn>=0.11.0
